Refactor hint logic into dedicated module
